drm/etnaviv: avoid deprecated timespec
struct timespec is being removed from the kernel because it often leads
to code that is not y2038-safe.

In the etnaviv driver, monotonic timestamps are used, which do not suffer
from overflow, but the usage of timespec here gets in the way of removing
the interface completely.

Pass down the user-supplied 64-bit value here rather than converting
it to an intermediate timespec to avoid the conversion.

The conversion is transparent for all regular CLOCK_MONOTONIC values,
but is a small change in behavior for excessively large values: the
existing code would treat e.g. tv_sec=0x100000000 the same as tv_sec=0
and not block, while the new code it would block for up to 2^31
seconds. The new behavior is more logical here, but if it causes problems,
the truncation can be put back.

Change-Id: I1623c8da25d966dcc242a9565bebcce0562b3fe6
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Git-commit: 531c0a438087a4108add1685667e5c54eb712171
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/arnd/playground.git
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>