soc: qcom: pil: Move away Removed DMA to ioremap API
PIL currently uses the removed DMA driver to allocate the
memory from the carveouts, and ensure that there is no kernel
mapping for these carveouts to save on vmalloc space. However,
the same effect can be achieved by reading the carveout
base address from the DT, and mapping pieces of the
carveout using ioremap().

However, peripherals, such as the GPU, may not use carveouts,
and as such, we should still be able to load them, so retain some
of the current behavior by migrating to dma_alloc_coherent()--to
ensure uncached mappings are used--for these peripherals. It is
important to note that this would not have same effect of
discarding the kernel mapping that exists for the region that
ends up being provided by dma_alloc_coherent().

Change-Id: I0bdbad0ab30a1b4ddb0b6bb0f48af8784f6865de
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Signed-off-by: Archit Saxena <archsaxe@codeaurora.org>