clk: qcom: clk-rcg2: Support update of RCG src/prediv
In the cases where RCG sources would support a slewing
parent and during a frequency switch if the new desired pre-div is
greater than the current pre-div then the RCG pre-div should be
updated before the slewing PLL source is updated.

There also could be a condition where the slewing parent alternate
outputs are used between different frequencies and in those conditions
if the new desired PLL frequency is greater than the current PLL
frequency the RCG source requires to be updated before the PLL rate
change.

Change-Id: I44ebca6f87ed9c99e5a324326995d18d3bd183d1
Signed-off-by: Taniya Das <tdas@codeaurora.org>