soc: qcom: silent mode monitoring driver
Kernel driver to monitor external GPIO input to indicate boot into
kernel silent mode or not. This driver sets up the GPIO and links it
up with an IRQ that is rising/falling edge triggered.
On the GPIO rising edge, the IRQ thread is used to clear the sysfs
in the PM framework that holds the state of the kernelsilentmode.

On the GPIO falling edge, the IRQ thread is used to set the sysfs
in the PM framework that holds the state of the kernelsilentmode.

Change-Id: I1f621b56f766f1ed0aeee739e8a500c39a6e4418
Signed-off-by: Manoj Vishwanathan <mvishwan@codeaurora.org>