usb: dwc3: gadget: Properly handle ClearFeature(halt)
Commit 1bfa09147282 ("usb: dwc3: gadget: Properly handle
ClearFeature(halt)") introduces stop active transfer to be done
as a part of clear feature. But for gsi as a part of clear
feature if stop active transfer is issued, as gsi uses hardware
path for transfers it cannot issue start active transfer which
leads to device sending not ready to the host while host trying
to send data which inturn leads to rmnet call failure. Fix this
by bailing out early without issuing stop active transfer in clear
feature for the interfaces that use hardware path for transfers.

Change-Id: I3b8b5762bc7eb24f4cff2c2e7147503ee5a3fa8f
Signed-off-by: Rohith Kollalsi <rkollals@codeaurora.org>
Signed-off-by: QU TAO <taoq@codeaurora.org>