drivers: thermal: cpu_voltage: Update the cooling device naming
The cooling device names has the cpu numbers ordered based on the cpu
with highest voltage. With voltage plan changes the cooling device names
change and this triggers change in config in userspace. To avoid the
name change, use the static order provided in devicetree to name the
cooling device.

Change-Id: I25f008d998ef05bd74f9dddeea2a410e00a24e38
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>