net: stmmac: persistent mac address
Add change to make sure addr_assign_type is set to persistent.
So that mac address can be persistent across reboot.
Currently eth_hw_addr_random() is called from stmmac &
it sets addr_assign_type to RANDOM. On LXC, system-udevd
check if the persistent is set or not if not it assign
the random Mac address.

Change-Id: Ib4abf5c48a398371f0a83049d6f045c8163410d0
Signed-off-by: Suraj Jaiswal <jsuraj@codeaurora.org>