power: qpnp-smblite: Add handshaking while enabling concurrency mode
In the use case where Audio playback is in progress and
charger is inserted, on enabling concurrency there is a
possibility of it to fail if the BOOST soft-start is not
complete. Avoid this scenario by polling on the BOOST_SW_DONE
bit which gets set after concurrency is successfully enabled
and then returning back to the caller.

Change-Id: I2e050c737c48413506b461c2184a94ceeda82119
Signed-off-by: Ashish Chavan <ashichav@codeaurora.org>