scsi: ufs-qcom: Move probe deferring for the 2nd UFS
For the platform having multiple ufs instance,we always
have 1st instance as the boot device. Due to async
scan of the device, there can be scenarios where
2nd ufs instance gets probed first and get enumerated.
Later when user space try to mount rootfs on that
partition we get permission error.

To handle this we are deferring the the probe of 2nd UFS.
Move the deferring of the 2nd UFS probe to the beginning
of the probe to save boot time and also to avoid changing the
resource states.

Change-Id: Ia0af2f77f399ecadfd9454facbd52a9594591f94
Signed-off-by: Veera Vegivada <quic_vvegivad@quicinc.com>