power: smblite: Update concurrency-mode failure handling
When there is a boost-fault during concurrency enablement,
the switcher goes back to BOOST mode when the concurrency
bit is disabled. In this path, wait for the SS_DONE to be
set before returning back from the call to ensure that the
BOOST is ready for the audio load.

While at it, use the BOOST_EN bit to check the state of the
BOOST instead of CHARGING_DISABLED_FROM_BOOST_BIT which may
be reset to 0 when concurrency is enabled hence causing the
unexpected QC3 increment to kick in.

Change-Id: Iea0aea6559e5a8a04f2acba917e6240f625cab8c
Signed-off-by: Ashish Chavan <ashichav@codeaurora.org>