iio: qcom-spmi-adc5: Disable interrupts in shutdown cb
On SKUs with MSM + Co-proc there can be possiblity that
because of Interrupts enabled from MSM side (SPMI) the exit
of LPM might not work because PBS gets locked because of
which the sequence to exit LPM might not get triggered.

Add shutdown callback hook and disable interrupts in shutdown
callback.

Change-Id: I613131ff33ee65a43135926db1dad66e59d78bf5
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>