power: supply: qbg: Set QBG FIFO-FULL INT as wakeable
Currently, QBG DATA_FULL_INT is not set as wakeable interrupt.
Hence in RBSC, device does not wake-up to process the interrupt
which causes QBG lock-up as INT_TEST_VAL reg is not cleared.

Add change to set DATA_FULL_INT as wakeable so that the device wakes
up from RBSC on DATA_FULL_INT to process QBG data.

Change-Id: Ibedf7f8f78deac20cac79b954242f2a95d6d2f72
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>