regulator: rpm-smd: Update mode voting support for PMIC5 LDOs
At present, only HPM and LPM modes are supported for PMIC5 LDO.
In some cases, clients may need to vote for retention mode on
them. To support this, update mode voting support for PMIC5 LDOs
by providing the option to specify a list of supported modes for
a PMIC5 LDO in devicetree instead of a single HPM min load value.

Change-Id: I0d058ef6763751e1aab669fcc8cbb91538b3bcb8
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>